+++ server/camcops_server/tasks/isaaqcommon.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 """
-camcops_server/tasks/isaaq.py
+camcops_server/tasks/isaaqcommon.py
 
 ===============================================================================