+++ tablet_qt/menu/setmenucpftadulteatingdisorders.cpp
@@ -33,7 +33,7 @@
 #include "tasks/gbogres.h"
 #include "tasks/gbogpc.h"
 #include "tasks/ided3d.h"
-#include "tasks/isaaq.h"
+#include "tasks/isaaq10.h"
 #include "tasks/isaaqed.h"
 #include "tasks/paradise24.h"
 #include "tasks/phq9.h"
@@ -84,7 +84,7 @@ void SetMenuCpftAdultEatingDisorders::makeItems()
         MenuItem(tr("Research")).setLabelOnly(),
         MAKE_TASK_MENU_ITEM(Chit::CHIT_TABLENAME, m_app),
         MAKE_TASK_MENU_ITEM(IDED3D::IDED3D_TABLENAME, m_app),
-        MAKE_TASK_MENU_ITEM(Isaaq::ISAAQ_TABLENAME, m_app),
+        MAKE_TASK_MENU_ITEM(Isaaq10::ISAAQ10_TABLENAME, m_app),
         MAKE_TASK_MENU_ITEM(IsaaqEd::ISAAQED_TABLENAME, m_app),
         MAKE_TASK_MENU_ITEM(Suppsp::SUPPSP_TABLENAME, m_app),
     };